Refactor storage controller
